Module Domains.Bitwise

module Make (Sub : Sig.BASE) : Sig.BASE with module Context = Sub.Context and type boolean = Sub.boolean

A domain that abstracts a bitvector, and remembers what bitwise operations have been applied to it. Can be a very efficient way to simplify sequences of bitwise operations.