Module Domains.Memory_domains

module Assert_false : sig ... end

module Block_smashing : sig ... end

module Convert : sig ... end

Context conversion procedures: pass through the values by just changing the context.

module Flexible_array_member : sig ... end

module Fully_expanded_finite_region : sig ... end

module Memory_domain : sig ... end

module Region_numeric_offset : sig ... end

Lifts a Scalar domain into a Memory_domain, which directly uses the Scalar as offsets in a finite, contiguous memory region (implemented using Fully_expanded_finite_region)

module Region_separation : sig ... end

module Region_suffix_tree : sig ... end

module Symbolic_boolean : sig ... end

module Typed_address : sig ... end

module Value_union_concatenation : sig ... end

module Wholify : sig ... end

module With_focusing : sig ... end