Module Symbolic_boolean.Make

Parameters

module B : AbstractBoolean

Signature

module Scalar := B.Scalar

include Sig.With_Boolean with module Context = B.Scalar.Context

module Context = B.Scalar.Context

type boolean

module Boolean : Datatype_sig.S with type t = boolean

val boolean_pretty : Context.t -> Format.formatter -> boolean -> unit

val serialize_boolean : Context.t -> boolean -> Context.t -> boolean -> 'a Context.in_acc -> (boolean, 'a) Context.result

val boolean_empty : Context.t -> boolean

Empty denotes that the concretization has no value (or it is the concrete value representing the absence of value). Note that this does not necessarily imply that some error occured; for instance the offset of an uninitialized pointer can be represented with empty. Emptyness testing is a simple way of communicating between domains.

val boolean_unknown : Context.t -> boolean

module Boolean_Forward : Operator.BOOLEAN_FORWARD with module Arity := Sig.Context_Arity_Forward(Context) and type boolean := boolean

val query_boolean : Context.t -> boolean -> Lattices.Quadrivalent.t

val lift : B.boolean -> boolean

val assume : Context.t -> boolean -> Context.t option

val satisfiable : Context.t -> boolean -> Smtbackend.Smtlib_sig.sat