Module type Memory_sig.WHOLE_MEMORY_DOMAIN

Like ADDRESS_AND_MAKE_MEMORY, but we can do bitvector operations on addresses. This module is not fully instantiated as module MEMORY still a functor waiting for parameter of module type BLOCK

module Scalar : Sig.BASE

module Offset : OFFSET with module Scalar = Scalar

module Address : FIXED_SIZE_VALUE_DOMAIN with module Scalar = Scalar

module Make_Memory (Block : BLOCK with module Scalar = Scalar and module Offset = Offset) : MEMORY with module Scalar = Scalar and module Address := Address and module Block := Block