Module type Memory_sig.WITH_ADDRESS

Abstraction of an address, pointing to a region in a C-like memory model. The type adress is named binary for historical reasons, and also to help sharing with bitvectors in whole.

include AADT

module Scalar : Sig.BASE

type boolean

type binary

module Binary : Datatype_sig.S with type t = binary

val binary_pretty : size:Units.In_bits.t -> Scalar.Context.t -> Format.formatter -> binary -> unit

val binary_empty : size:Units.In_bits.t -> Scalar.Context.t -> binary

val satisfiable : Scalar.Context.t -> boolean -> Smtbackend.Smtlib_sig.sat

val serialize : widens:bool -> size:Units.In_bits.t -> Scalar.Context.t -> binary -> Scalar.Context.t -> binary -> 'a Scalar.Context.in_acc -> (binary, 'a) Scalar.Context.result

Serialize functions also return their de-serializer.

val bchoose : size:Units.In_bits.t -> Operator.Choice.t -> Scalar.Context.t -> binary -> binary

val bindex : size:Units.In_bits.t -> int -> Scalar.Context.t -> binary -> Scalar.binary -> binary

bindex ~size k ctx address expr is address + k * expr

val bisub : size:Units.In_bits.t -> Scalar.Context.t -> binary -> binary -> Scalar.binary

pointer substraction.

val bshift : size:Units.In_bits.t -> offset:int -> max:int option -> Scalar.Context.t -> binary -> binary

bshift ~size ~offset ~max ctx address is address + offset. In addition, further shifts/index are bounded by max.

val ble : size:Units.In_bits.t -> Scalar.Context.t -> binary -> binary -> boolean

Pointer comparison.

val beq : size:Units.In_bits.t -> Scalar.Context.t -> binary -> binary -> boolean

val binary_unknown : size:Units.In_bits.t -> Scalar.Context.t -> binary

val binary_unknown_typed : size:Units.In_bits.t -> Scalar.Context.t -> Codex_types.TypedC.typ -> binary

val check_type : size:Units.In_bits.t -> Scalar.Context.t -> Codex_types.TypedC.typ -> binary -> binary Codex_types.Type_check_tree.t

Type check a binary value

val binary2scalar_binary : size:Units.In_bits.t -> Scalar.Context.t -> binary -> Scalar.binary

val type_of : size:Units.In_bits.t -> Scalar.Context.t -> binary -> Codex_types.TypedC.typ option

Returns the type of a function, or None if unknown.

val within_bounds : size:Units.In_bits.t -> Scalar.Context.t -> binary -> boolean

Check that all array indices operation are within bound.

module Query : sig ... end