Module WITH_BLOCK_FORWARD.Block_Forward

val sizeof : (block, offset) Sig.Context_Arity_Forward(Context).ar1

Size of a block in bytes

val concat : (block, block, block) Sig.Context_Arity_Forward(Context).ar2

Concatenates two blocks

val load : size:Units.In_bits.t -> (block, offset, value) Sig.Context_Arity_Forward(Context).ar2

Loads (extracts) a value of a fixed size at a given index from a block

val store : size:Units.In_bits.t -> (block, offset, value, block) Sig.Context_Arity_Forward(Context).ar3

Stores (writes) a fixed size value of a given index in a block

val binary_to_block : size:Units.In_bits.t -> (value, block) Sig.Context_Arity_Forward(Context).ar1

Converts a fixed size value to a block