Module type Sig.Ext

These are functions that can be implemented using the base signatures. See Domain_extend for instantiation.

module Context : Context

type boolean

val imperative_assume : Context.t -> boolean -> unit

Because the transfer functions imperatively change the context, they cannot use assume, that returns a new context. Temporarily, we provide this instead (it should be applied only to fresh symbolic variables and not modify the set of valuations of the other symbolic variables. In particular, the condition must never make the context bottom).

The good long-term solution would be to make every transfer function return a new Context.t option, viewing the context as some state monad.